Crate sp_block_producer_metadata

Block Producer Metadata Primitives

This crate contains primitive types and logic used by the Block Producer Metadata feature of the Partner Chain Toolkit. This feature allows Partner Chain block producers to save information about themselves on-chain. The format of this metadata is left generic for each Partner Chain builder to define. Metadata is associated with the cross-chain public key of the block producer.

Currently the only code defined in this crate is the MetadataSignedMessage type describing the message that is signed and submitted by the block producer together with each change in their on-chain metadata to prove that they are the owner of the public keys associated with the metadata.

Structs

MetadataSignedMessage

Message signed to authorize modification of a block producer’s on-chain metadata

Traits

BlockProducerMetadataApi

Runtime API for accessing metadata of block producers